Update README, revert the changes

+++ src/main/java/appeng/client/render/cablebus/CableBuilder.java
@@ -333,9 +333,9 @@ class CableBuilder {
         cubeBuilder.setTexture(texture);
 
         TextureAtlasSprite oddChannel = this.smartCableTextures
-                .getOddTextureForChannels(channels / (AEConfig.instance().getNormalChannelCapacity() / 8));
+                .getOddTextureForChannels((int) (channels / (AEConfig.instance().getNormalChannelCapacity() / 8)));
         TextureAtlasSprite evenChannel = this.smartCableTextures
-                .getEvenTextureForChannels(channels / (AEConfig.instance().getNormalChannelCapacity() / 8));
+                .getEvenTextureForChannels((int) (channels / (AEConfig.instance().getNormalChannelCapacity() / 8)));
 
         // Thin part of connector, here to prevent the color of the channels
         // from leaking into it
@@ -385,9 +385,9 @@ class CableBuilder {
         addStraightCoveredCableSizedCube(facing, cubeBuilder);
 
         TextureAtlasSprite oddChannel = this.smartCableTextures
-                .getOddTextureForChannels(channels / (AEConfig.instance().getNormalChannelCapacity() / 8));
+                .getOddTextureForChannels((int) (channels / (AEConfig.instance().getNormalChannelCapacity() / 8)));
         TextureAtlasSprite evenChannel = this.smartCableTextures
-                .getEvenTextureForChannels(channels / (AEConfig.instance().getNormalChannelCapacity() / 8));
+                .getEvenTextureForChannels((int) (channels / (AEConfig.instance().getNormalChannelCapacity() / 8)));
 
         // Render the channel indicators brightly lit at night
         cubeBuilder.setRenderFullBright(true);
@@ -417,9 +417,9 @@ class CableBuilder {
         addCoveredCableSizedCube(facing, distanceFromEdge, cubeBuilder);
 
         TextureAtlasSprite oddChannel = this.smartCableTextures
-                .getOddTextureForChannels(channels / (AEConfig.instance().getNormalChannelCapacity() / 8));
+                .getOddTextureForChannels((int) (channels / (AEConfig.instance().getNormalChannelCapacity() / 8)));
         TextureAtlasSprite evenChannel = this.smartCableTextures
-                .getEvenTextureForChannels(channels / (AEConfig.instance().getNormalChannelCapacity() / 8));
+                .getEvenTextureForChannels((int) (channels / (AEConfig.instance().getNormalChannelCapacity() / 8)));
 
         // Render the channel indicators brightly lit at night
         cubeBuilder.setRenderFullBright(true);
@@ -482,7 +482,7 @@ class CableBuilder {
         addDenseCableSizedCube(facing, cubeBuilder);
 
         // Dense cables show used channels in groups of 4, rounded up
-        channels = (channels + 3) / 4 / (AEConfig.instance().getDenseChannelCapacity() / 32);
+        channels = (int) ((channels + 3) / 4) / (AEConfig.instance().getDenseChannelCapacity() / 32);
 
         TextureAtlasSprite oddChannel = this.smartCableTextures.getOddTextureForDenseChannels(channels);
         TextureAtlasSprite evenChannel = this.smartCableTextures.getEvenTextureForDenseChannels(channels);
@@ -527,7 +527,7 @@ class CableBuilder {
         addStraightDenseCableSizedCube(facing, cubeBuilder);
 
         // Dense cables show used channels in groups of 4, rounded up
-        channels = (channels + 3) / 4 / (AEConfig.instance().getDenseChannelCapacity() / 32);
+        channels = (int) ((channels + 3) / 4) / (AEConfig.instance().getDenseChannelCapacity() / 32);
 
         TextureAtlasSprite oddChannel = this.smartCableTextures.getOddTextureForDenseChannels(channels);
         TextureAtlasSprite evenChannel = this.smartCableTextures.getEvenTextureForDenseChannels(channels);
